--- programs/helium-entity-manager/src/instructions/metadata/mod.rs
@@ -1,9 +1,0 @@
-pub mod assert_elevation_v0;
-pub mod assert_gain_v0;
-pub mod assert_location_v0;
-pub mod change_metadata_v0;
-
-pub use assert_elevation_v0::*;
-pub use assert_gain_v0::*;
-pub use assert_location_v0::*;
-pub use change_metadata_v0::*;
