+++ programs/helium-entity-manager/src/instructions/mod.rs
@@ -1,9 +1,9 @@
+pub mod change_metadata_v0;
 pub mod initialize_hotspot_config_v0;
 pub mod initialize_hotspot_issuer_v0;
 pub mod issue_hotspot_v0;
-pub mod metadata;
 
+pub use change_metadata_v0::*;
 pub use initialize_hotspot_config_v0::*;
 pub use initialize_hotspot_issuer_v0::*;
 pub use issue_hotspot_v0::*;
-pub use metadata::*;
